+++ services/api-server/src/simcore_service_api_server/api/routes/solvers_jobs_getters.py
@@ -21,7 +21,7 @@ from servicelib.fastapi.requests_decorators import cancel_on_disconnect
 from servicelib.logging_utils import log_context
 from starlette.background import BackgroundTask
 
-from ...exceptions.custom_exceptions import InsufficientCreditsError, MissingWalletError
+from ...exceptions.custom import InsufficientCreditsError, MissingWalletError
 from ...exceptions.services_exception_handlers import (
     DEFAULT_BACKEND_SERVICE_STATUS_CODES,
 )
