+++ services/api-server/src/simcore_service_api_server/exceptions/_http_error_handlers.py
