--- services/api-server/src/simcore_service_api_server/exceptions/log_exceptions.py
@@ -1,13 +1,0 @@
-from .base_exceptions import ApiServerBaseError
-
-
-class LogDistributionBaseError(ApiServerBaseError):
-    pass
-
-
-class LogStreamerNotRegisteredError(LogDistributionBaseError):
-    pass
-
-
-class LogStreamerRegistionConflictError(LogDistributionBaseError):
-    pass
