+++ services/api-server/src/simcore_service_api_server/exceptions/log_streaming.py
@@ -1,0 +1,13 @@
+from ._base import ApiServerBaseError
+
+
+class LogStreamingBaseError(ApiServerBaseError):
+    pass
+
+
+class LogStreamerNotRegisteredError(LogStreamingBaseError):
+    pass
+
+
+class LogStreamerRegistionConflictError(LogStreamingBaseError):
+    pass
